move pr detection into separate file
